refactor: :hammer: removed unused files, refactor in routes
