+++ app/src/main/java/com/rodrigo/deeplarva/routes/service/PredictionService.kt
@@ -22,7 +22,7 @@ import com.rodrigo.deeplarva.modules.prediction.BackgroundTaskPredict
 import com.rodrigo.deeplarva.infraestructure.services.BoxDetectionServices
 import com.rodrigo.deeplarva.infraestructure.services.PicturesServices
 import com.rodrigo.deeplarva.routes.activity.PicturesActivity
-import com.rodrigo.deeplarva.routes.broadcast.PredictionBroadcastSender
+import com.rodrigo.deeplarva.routes.service.broadcast.PredictionBroadcastSender
 
 class PredictionService: Service() {
     private val TAG = "DEEP_LARVA::PredictionService"
