+++ app/src/main/java/com/rodrigo/deeplarva/routes/service/binder/IBoundService.kt
@@ -1,4 +1,4 @@
-package com.rodrigo.deeplarva.routes.binder
+package com.rodrigo.deeplarva.routes.service.binder
 
 interface IBoundService {
     fun onBindToService()
