+++ app/src/main/java/com/rodrigo/deeplarva/routes/service/broadcast/PredictionBroadcastSender.kt
@@ -1,4 +1,4 @@
-package com.rodrigo.deeplarva.routes.broadcast
+package com.rodrigo.deeplarva.routes.service.broadcast
 
 import android.app.Service
 import android.content.Intent
