--- app/src/main/res/drawable/border_background.xml
@@ -1,18 +1,0 @@
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <!-- Background with rounded corners -->
-    <item>
-        <shape android:shape="rectangle">
-            <corners android:radius="0dp" />
-            <solid android:color="@android:color/transparent" />
-        </shape>
-    </item>
-    <!-- Border -->
-    <item>
-        <shape android:shape="rectangle">
-            <corners android:radius="0dp" />
-            <stroke
-                android:width="1.5dp"
-                android:color="@color/gray500" />
-        </shape>
-    </item>
-</layer-list>
