+++ app/src/main/res/drawable/ic_launcher_background.xml
@@ -5,7 +5,7 @@
     android:viewportHeight="108"
     android:viewportWidth="108"
     xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#3DDC84"
+    <path android:fillColor="#FFFFFFFF"
           android:pathData="M0,0h108v108h-108z"/>
     <path android:fillColor="#00000000" android:pathData="M9,0L9,108"
           android:strokeColor="#33FFFFFF" android:strokeWidth="0.8"/>
