--- app/src/main/res/drawable/ic_outer_circle.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Copyright 2020 The Android Open Source Project
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     https://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_pressed="true">
-        <shape android:shape="oval">
-            <stroke
-                android:width="@dimen/stroke_small"
-                android:color="#BDC3C7" />
-            <size
-                android:width="@dimen/round_button_medium"
-                android:height="@dimen/round_button_medium" />
-        </shape>
-    </item>
-    <item android:state_focused="true">
-        <shape android:shape="oval">
-            <stroke
-                android:width="@dimen/stroke_small"
-                android:color="@color/orange" />
-            <size
-                android:width="@dimen/round_button_medium"
-                android:height="@dimen/round_button_medium" />
-        </shape>
-    </item>
-    <item>
-        <shape android:shape="oval">
-            <stroke
-                android:width="@dimen/stroke_small"
-                android:color="#718792" />
-            <size
-                android:width="@dimen/round_button_medium"
-                android:height="@dimen/round_button_medium" />
-        </shape>
-    </item>
-</selector>
