+++ app/src/main/res/drawable/rounded_corner_camera_command.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/gray100"/> <!-- Background color of the RelativeLayout -->
+    <corners
+        android:topLeftRadius="4dp"
+        android:bottomLeftRadius="4dp"/>
+</shape>
