--- app/src/main/res/drawable/rounded_corners.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="#F4F4F9"/> <!-- Background color of the RelativeLayout -->
-    <corners
-        android:topLeftRadius="20dp"
-        android:topRightRadius="20dp"
-        android:bottomRightRadius="20dp"
-        android:bottomLeftRadius="20dp"/>
-</shape>
