--- app/src/main/res/drawable/rounded_corners_top.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="#79767D"/> <!-- Background color of the RelativeLayout -->
-    <corners
-        android:topLeftRadius="20dp"
-        android:topRightRadius="20dp"/>
-</shape>
